Include filename in cache key
When an ES6 file is renamed, we need to recompile it, even if the
contents hasn't changed. This is because the module's path has changed,
and depending on the module loader, the compiled output will need to
reflect this.

For example, using SystemJS, if you rename an empty file from app.es6
to app1.es6, the output needs to change from:

System.register("app", [], function (_export) {
  ... rest of output
});

To:

System.register("app1", [], function (_export) {
  ... rest of output
});